Extract server commands to mixin
