+++ fakeredis/commands_mixins/__init__.py
@@ -3,4 +3,4 @@ from .hash_mixin import HashCommandsMixin
 from .scripting_mixin import ScriptingCommandsMixin
 from .connection_mixin import ConnectionCommandsMixin
 from .list_mixin import ListCommandsMixin
-
+from .server_mixin import ServerCommandsMixin
