+++ test/test_fakeredis6.py
@@ -2,7 +2,7 @@ import math
 import os
 import threading
 from collections import OrderedDict
-from datetime import datetime, timedelta
+from datetime import timedelta
 from queue import Queue
 from time import sleep
 
@@ -33,26 +33,6 @@ def zincrby(r, key, amount, value):
 def test_large_command(r):
     r.set('foo', 'bar' * 10000)
     assert r.get('foo') == b'bar' * 10000
-
-
-def test_dbsize(r):
-    assert r.dbsize() == 0
-    r.set('foo', 'bar')
-    r.set('bar', 'foo')
-    assert r.dbsize() == 2
-
-
-def test_flushdb(r):
-    r.set('foo', 'bar')
-    assert r.keys() == [b'foo']
-    assert r.flushdb() is True
-    assert r.keys() == []
-
-
-def test_set_float_value(r):
-    x = 1.23456789123456789
-    r.set('foo', x)
-    assert float(r.get('foo')) == x
 
 
 def test_saving_non_ascii_chars_as_value(r):
@@ -2273,11 +2253,6 @@ def test_key_patterns(r):
     assert sorted(r.keys()) == [b'four', b'one', b'three', b'two']
 
 
-def test_ping(r):
-    assert r.ping()
-    assert raw_command(r, 'ping', 'test') == b'test'
-
-
 @testtools.run_test_if_redispy_ver('above', '3')
 def test_ping_pubsub(r):
     p = r.pubsub()
@@ -2287,62 +2262,6 @@ def test_ping_pubsub(r):
     assert p.parse_response() == [b'pong', b'']
     p.ping('test')
     assert p.parse_response() == [b'pong', b'test']
-
-
-@testtools.run_test_if_redispy_ver('above', '3')
-def test_swapdb(r, create_redis):
-    r1 = create_redis(1)
-    r.set('foo', 'abc')
-    r.set('bar', 'xyz')
-    r1.set('foo', 'foo')
-    r1.set('baz', 'baz')
-    assert r.swapdb(0, 1)
-    assert r.get('foo') == b'foo'
-    assert r.get('bar') is None
-    assert r.get('baz') == b'baz'
-    assert r1.get('foo') == b'abc'
-    assert r1.get('bar') == b'xyz'
-    assert r1.get('baz') is None
-
-
-@testtools.run_test_if_redispy_ver('above', '3')
-def test_swapdb_same_db(r):
-    assert r.swapdb(1, 1)
-
-
-def test_save(r):
-    assert r.save()
-
-
-def test_bgsave(r):
-    assert r.bgsave()
-    with pytest.raises(ResponseError):
-        r.execute_command('BGSAVE', 'SCHEDULE', 'FOO')
-    with pytest.raises(ResponseError):
-        r.execute_command('BGSAVE', 'FOO')
-
-
-def test_lastsave(r):
-    assert isinstance(r.lastsave(), datetime)
-
-
-@pytest.mark.slow
-def test_bgsave_timestamp_update(r):
-    early_timestamp = r.lastsave()
-    sleep(1)
-    assert r.bgsave()
-    sleep(1)
-    late_timestamp = r.lastsave()
-    assert early_timestamp < late_timestamp
-
-
-@pytest.mark.slow
-def test_save_timestamp_update(r):
-    early_timestamp = r.lastsave()
-    sleep(1)
-    assert r.save()
-    late_timestamp = r.lastsave()
-    assert early_timestamp < late_timestamp
 
 
 @pytest.mark.slow
